requirements.txt
Django==5.1.2
djangorestframework==3.14.0
coreapi==2.0.2
coreschema==0.0.4
uritemplate==4.1.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
dj-database-url==2.1.0
django-filter==24.2 